+++ game/game_libs/dlls/buttons.cpp
@@ -158,7 +158,7 @@ void CMultiSource::Spawn()
 
 	pev->solid = SOLID_NOT;
 	pev->movetype = MOVETYPE_NONE;
-	pev->nextthink = gpGlobals->time + 0.1;
+	pev->nextthink = gpGlobals->time + 0.1f;
 	pev->spawnflags |= SF_MULTI_INIT;  // Until it's initialized
 	SetThink(&CMultiSource::Register);
 }
@@ -379,22 +379,22 @@ void CBaseButton::KeyValue(KeyValueData* pkvd)
 	}
 	else if ( FStrEq(pkvd->szKeyName, "locked_sound") )
 	{
-		m_bLockedSound = atoi(pkvd->szValue);
+		m_bLockedSound = static_cast<BYTE>(atoi(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if ( FStrEq(pkvd->szKeyName, "locked_sentence") )
 	{
-		m_bLockedSentence = atoi(pkvd->szValue);
+		m_bLockedSentence = static_cast<BYTE>(atoi(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if ( FStrEq(pkvd->szKeyName, "unlocked_sound") )
 	{
-		m_bUnlockedSound = atoi(pkvd->szValue);
+		m_bUnlockedSound = static_cast<BYTE>(atoi(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if ( FStrEq(pkvd->szKeyName, "unlocked_sentence") )
 	{
-		m_bUnlockedSentence = atoi(pkvd->szValue);
+		m_bUnlockedSentence = static_cast<BYTE>(atoi(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if ( FStrEq(pkvd->szKeyName, "sounds") )
@@ -474,7 +474,7 @@ void CBaseButton::Spawn()
 	if ( FBitSet(pev->spawnflags, SF_BUTTON_SPARK_IF_OFF) )  // this button should spark in OFF state
 	{
 		SetThink(&CBaseButton::ButtonSpark);
-		pev->nextthink = gpGlobals->time + 0.5;  // no hurry, make sure everything else spawns
+		pev->nextthink = gpGlobals->time + 0.5f;  // no hurry, make sure everything else spawns
 	}
 
 	SetMovedir(pev);
@@ -501,8 +501,8 @@ void CBaseButton::Spawn()
 	// Subtract 2 from size because the engine expands bboxes by 1 in all directions making the size too big
 	m_vecPosition2 = m_vecPosition1 +
 		(pev->movedir *
-		 (fabs(pev->movedir.x * (pev->size.x - 2)) + fabs(pev->movedir.y * (pev->size.y - 2)) +
-		  fabs(pev->movedir.z * (pev->size.z - 2)) - m_flLip));
+		 (fabsf(pev->movedir.x * (pev->size.x - 2)) + fabsf(pev->movedir.y * (pev->size.y - 2)) +
+		  fabsf(pev->movedir.z * (pev->size.z - 2)) - m_flLip));
 
 	// Is this a non-moving button?
 	if ( ((m_vecPosition2 - m_vecPosition1).Length() < 1) || (pev->spawnflags & SF_BUTTON_DONTMOVE) )
@@ -611,7 +611,7 @@ void DoSpark(entvars_t* pev, const Vector& location)
 	Vector tmp = location + pev->size * 0.5;
 	UTIL_Sparks(tmp);
 
-	float flVolume = RANDOM_FLOAT(0.25, 0.75) * 0.4;  // random volume range
+	float flVolume = RANDOM_FLOAT(0.25f, 0.75f) * 0.4f;  // random volume range
 	switch ( (int)(RANDOM_FLOAT(0, 1) * 6) )
 	{
 		case 0:
@@ -638,7 +638,7 @@ void DoSpark(entvars_t* pev, const Vector& location)
 void CBaseButton::ButtonSpark(void)
 {
 	SetThink(&CBaseButton::ButtonSpark);
-	pev->nextthink = gpGlobals->time + 0.1 + RANDOM_FLOAT(0, 1.5);  // spark again at random interval
+	pev->nextthink = gpGlobals->time + 0.1f + RANDOM_FLOAT(0.0f, 1.5f);  // spark again at random interval
 
 	DoSpark(pev, pev->mins);
 }
@@ -852,7 +852,7 @@ void CBaseButton::ButtonBackHome(void)
 	if ( FBitSet(pev->spawnflags, SF_BUTTON_SPARK_IF_OFF) )
 	{
 		SetThink(&CBaseButton::ButtonSpark);
-		pev->nextthink = gpGlobals->time + 0.5;  // no hurry.
+		pev->nextthink = gpGlobals->time + 0.5f;  // no hurry.
 	}
 }
 
@@ -1022,7 +1022,7 @@ void CMomentaryRotButton::KeyValue(KeyValueData* pkvd)
 {
 	if ( FStrEq(pkvd->szKeyName, "returnspeed") )
 	{
-		m_returnSpeed = atof(pkvd->szValue);
+		m_returnSpeed = static_cast<float>(atof(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if ( FStrEq(pkvd->szKeyName, "sounds") )
@@ -1090,7 +1090,7 @@ void CMomentaryRotButton::UpdateSelf(float value)
 	}
 	m_lastUsed = 1;
 
-	pev->nextthink = pev->ltime + 0.1;
+	pev->nextthink = pev->ltime + 0.1f;
 	if ( m_direction > 0 && value >= 1.0 )
 	{
 		pev->avelocity = g_vecZero;
@@ -1110,9 +1110,9 @@ void CMomentaryRotButton::UpdateSelf(float value)
 	// HACKHACK -- If we're going slow, we'll get multiple player packets per frame, bump nexthink on each one to avoid
 	// stalling
 	if ( pev->nextthink < pev->ltime )
-		pev->nextthink = pev->ltime + 0.1;
+		pev->nextthink = pev->ltime + 0.1f;
 	else
-		pev->nextthink += 0.1;
+		pev->nextthink += 0.1f;
 
 	pev->avelocity = m_direction * pev->speed * pev->movedir;
 	SetThink(&CMomentaryRotButton::Off);
@@ -1144,7 +1144,7 @@ void CMomentaryRotButton::Off(void)
 	if ( FBitSet(pev->spawnflags, SF_PENDULUM_AUTO_RETURN) && m_returnSpeed > 0 )
 	{
 		SetThink(&CMomentaryRotButton::Return);
-		pev->nextthink = pev->ltime + 0.1;
+		pev->nextthink = pev->ltime + 0.1f;
 		m_direction = -1;
 	}
 	else
@@ -1172,7 +1172,7 @@ void CMomentaryRotButton::UpdateSelfReturn(float value)
 	else
 	{
 		pev->avelocity = -m_returnSpeed * pev->movedir;
-		pev->nextthink = pev->ltime + 0.1;
+		pev->nextthink = pev->ltime + 0.1f;
 	}
 }
 
@@ -1224,7 +1224,7 @@ void CEnvSpark::Spawn(void)
 	else
 		SetThink(&CEnvSpark::SparkThink);
 
-	pev->nextthink = gpGlobals->time + 0.1 + RANDOM_FLOAT(0, 1.5);
+	pev->nextthink = gpGlobals->time + 0.1f + RANDOM_FLOAT(0.0f, 1.5f);
 
 	if ( m_flDelay <= 0 )
 		m_flDelay = 1.5;
@@ -1246,7 +1246,7 @@ void CEnvSpark::KeyValue(KeyValueData* pkvd)
 {
 	if ( FStrEq(pkvd->szKeyName, "MaxDelay") )
 	{
-		m_flDelay = atof(pkvd->szValue);
+		m_flDelay = static_cast<float>(atof(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if (
@@ -1260,7 +1260,7 @@ void CEnvSpark::KeyValue(KeyValueData* pkvd)
 
 void EXPORT CEnvSpark::SparkThink(void)
 {
-	pev->nextthink = gpGlobals->time + 0.1 + RANDOM_FLOAT(0, m_flDelay);
+	pev->nextthink = gpGlobals->time + 0.1f + RANDOM_FLOAT(0.0f, m_flDelay);
 	DoSpark(pev, pev->origin);
 }
 
@@ -1268,7 +1268,7 @@ void EXPORT CEnvSpark::SparkStart(CBaseEntity*, CBaseEntity*, USE_TYPE, float)
 {
 	SetUse(&CEnvSpark::SparkStop);
 	SetThink(&CEnvSpark::SparkThink);
-	pev->nextthink = gpGlobals->time + 0.1 + RANDOM_FLOAT(0, m_flDelay);
+	pev->nextthink = gpGlobals->time + 0.1f + RANDOM_FLOAT(0.0f, m_flDelay);
 }
 
 void EXPORT CEnvSpark::SparkStop(CBaseEntity*, CBaseEntity*, USE_TYPE, float)
